requirements.txt
-e git+https://github.com/django/django.git@stable/1.6.x#egg=django
South>=0.8.2
djangorestframework>=2.3.7
-e git+https://github.com/karolmajta/filthy-tricks.git@9b00823b3c783b87b2386871b74aa25430d6e2b4#egg=filthy-tricks
